+++ crates/remote/src/grpc_remote_client.rs
@@ -7,40 +7,95 @@ use bazel_remote_apis::build::bazel::remote::execution::v2::{
     action_cache_client::ActionCacheClient, batch_update_blobs_request,
     capabilities_client::CapabilitiesClient,
     content_addressable_storage_client::ContentAddressableStorageClient, digest_function,
-    ActionResult, BatchReadBlobsRequest, BatchUpdateBlobsRequest, Digest, GetActionResultRequest,
-    GetCapabilitiesRequest, ServerCapabilities, UpdateActionResultRequest,
+    ActionResult, BatchReadBlobsRequest, BatchUpdateBlobsRequest, Digest, FindMissingBlobsRequest,
+    GetActionResultRequest, GetCapabilitiesRequest, ServerCapabilities, UpdateActionResultRequest,
 };
+use miette::IntoDiagnostic;
 use moon_common::color;
-use moon_config::{RemoteCompression, RemoteConfig};
-use std::{error::Error, path::Path};
+use moon_config::RemoteConfig;
+use std::{env, path::Path, str::FromStr};
 use tonic::{
+    metadata::{KeyAndValueRef, MetadataKey, MetadataMap, MetadataValue},
     transport::{Channel, Endpoint},
-    Code,
+    Code, Request, Status,
 };
 use tracing::{trace, warn};
 
 fn map_transport_error(error: tonic::transport::Error) -> RemoteError {
+    // dbg!(&error);
     RemoteError::GrpcConnectFailed {
         error: Box::new(error),
     }
 }
 
 fn map_status_error(error: tonic::Status) -> RemoteError {
-    match error.source() {
-        Some(src) => RemoteError::GrpcCallFailedViaSource {
-            error: src.to_string(),
-        },
-        None => RemoteError::GrpcCallFailed {
-            error: Box::new(error),
-        },
+    // dbg!(&error);
+    RemoteError::GrpcCallFailed {
+        error: Box::new(error),
     }
 }
 
 #[derive(Default)]
 pub struct GrpcRemoteClient {
     channel: Option<Channel>,
-    compression: RemoteCompression,
-    instance_name: String,
+    config: RemoteConfig,
+    headers: MetadataMap,
+}
+
+impl GrpcRemoteClient {
+    fn extract_headers(&mut self) -> miette::Result<bool> {
+        let mut enabled = true;
+
+        if let Some(auth) = &self.config.auth {
+            for (key, value) in &auth.headers {
+                self.headers.insert(
+                    MetadataKey::from_str(key).into_diagnostic()?,
+                    MetadataValue::from_str(value).into_diagnostic()?,
+                );
+            }
+
+            if let Some(token_name) = &auth.token {
+                let token = env::var(token_name).unwrap_or_default();
+
+                if token.is_empty() {
+                    enabled = false;
+
+                    warn!(
+                        "Auth token {} does not exist, unable to authorize for remote service",
+                        color::property(token_name)
+                    );
+                } else {
+                    self.headers.insert(
+                        MetadataKey::from_str("Authorization").into_diagnostic()?,
+                        MetadataValue::from_str(&format!("Bearer {token}")).into_diagnostic()?,
+                    );
+                }
+            }
+        }
+
+        Ok(enabled)
+    }
+
+    fn inject_auth_headers(&self, mut req: Request<()>) -> Result<Request<()>, Status> {
+        if self.headers.is_empty() {
+            return Ok(req);
+        }
+
+        let headers = req.metadata_mut();
+
+        for entry in self.headers.iter() {
+            match entry {
+                KeyAndValueRef::Ascii(key, value) => {
+                    headers.insert(key.clone(), value.clone());
+                }
+                KeyAndValueRef::Binary(key, value) => {
+                    headers.insert_bin(key.clone(), value.clone());
+                }
+            };
+        }
+
+        Ok(req)
+    }
 }
 
 #[async_trait::async_trait]
@@ -49,7 +104,7 @@ impl RemoteClient for GrpcRemoteClient {
         &mut self,
         config: &RemoteConfig,
         workspace_root: &Path,
-    ) -> miette::Result<()> {
+    ) -> miette::Result<bool> {
         let host = &config.host;
 
         trace!(
@@ -60,12 +115,22 @@ impl RemoteClient for GrpcRemoteClient {
                 "(with mTLS)"
             } else if config.tls.is_some() {
                 "(with TLS)"
+            } else if config.is_bearer_auth() {
+                "(with auth)"
             } else {
                 "(insecure)"
             }
         );
 
-        let mut endpoint = Endpoint::from_shared(host.to_owned())
+        // Although we use a grpc(s) protocol for the host,
+        // tonic only supports http(s), so change it
+        let url = if let Some(suffix) = host.strip_prefix("grpc") {
+            format!("http{suffix}")
+        } else {
+            host.to_owned()
+        };
+
+        let mut endpoint = Endpoint::from_shared(url)
             .map_err(map_transport_error)?
             .user_agent("moon")
             .map_err(map_transport_error)?
@@ -78,6 +143,10 @@ impl RemoteClient for GrpcRemoteClient {
         } else if let Some(tls) = &config.tls {
             endpoint = endpoint
                 .tls_config(create_tls_config(tls, workspace_root)?)
+                .map_err(map_transport_error)?;
+        } else if config.is_secure_protocol() {
+            endpoint = endpoint
+                .tls_config(create_native_tls_config()?)
                 .map_err(map_transport_error)?;
         }
 
@@ -92,22 +161,32 @@ impl RemoteClient for GrpcRemoteClient {
             );
         }
 
-        self.channel = Some(endpoint.connect().await.map_err(map_transport_error)?);
-        self.compression = config.cache.compression;
-        self.instance_name = config.cache.instance_name.clone();
+        self.config = config.to_owned();
+        let enabled = self.extract_headers()?;
 
-        Ok(())
+        // We can't inject auth headers into this initial connection,
+        // so defer the connection until a client is used
+        if self.config.is_bearer_auth() {
+            self.channel = Some(endpoint.connect_lazy());
+        } else {
+            self.channel = Some(endpoint.connect().await.map_err(map_transport_error)?);
+        }
+
+        Ok(enabled)
     }
 
     // https://github.com/bazelbuild/remote-apis/blob/main/build/bazel/remote/execution/v2/remote_execution.proto#L452
     async fn load_capabilities(&self) -> miette::Result<ServerCapabilities> {
-        let mut client = CapabilitiesClient::new(self.channel.clone().unwrap());
+        let mut client =
+            CapabilitiesClient::with_interceptor(self.channel.clone().unwrap(), |req| {
+                self.inject_auth_headers(req)
+            });
 
         trace!("Loading remote execution API capabilities from gRPC server");
 
         let response = client
             .get_capabilities(GetCapabilitiesRequest {
-                instance_name: self.instance_name.clone(),
+                instance_name: self.config.cache.instance_name.clone(),
             })
             .await
             .map_err(map_status_error)?;
@@ -117,13 +196,16 @@ impl RemoteClient for GrpcRemoteClient {
 
     // https://github.com/bazelbuild/remote-apis/blob/main/build/bazel/remote/execution/v2/remote_execution.proto#L170
     async fn get_action_result(&self, digest: &Digest) -> miette::Result<Option<ActionResult>> {
-        let mut client = ActionCacheClient::new(self.channel.clone().unwrap());
+        let mut client =
+            ActionCacheClient::with_interceptor(self.channel.clone().unwrap(), |req| {
+                self.inject_auth_headers(req)
+            });
 
         trace!(hash = &digest.hash, "Checking for a cached action result");
 
         match client
             .get_action_result(GetActionResultRequest {
-                instance_name: self.instance_name.clone(),
+                instance_name: self.config.cache.instance_name.clone(),
                 action_digest: Some(digest.to_owned()),
                 inline_stderr: true,
                 inline_stdout: true,
@@ -151,7 +233,19 @@ impl RemoteClient for GrpcRemoteClient {
                     trace!(hash = &digest.hash, "Cache miss on action result");
 
                     Ok(None)
+                }
+                // If we hit an out of range error, the payload is larger than the grpc
+                // limit, and will fail the entire pipeline. Instead of letting that
+                // happen, let's just do a cache miss instead...
+                else if matches!(status.code(), Code::OutOfRange) {
+                    trace!(
+                        hash = &digest.hash,
+                        "Cache miss because the expected payload is too large"
+                    );
+
+                    Ok(None)
                 } else {
+                    // dbg!("get_action_result", digest);
                     Err(map_status_error(status).into())
                 }
             }
@@ -164,7 +258,10 @@ impl RemoteClient for GrpcRemoteClient {
         digest: &Digest,
         result: ActionResult,
     ) -> miette::Result<Option<ActionResult>> {
-        let mut client = ActionCacheClient::new(self.channel.clone().unwrap());
+        let mut client =
+            ActionCacheClient::with_interceptor(self.channel.clone().unwrap(), |req| {
+                self.inject_auth_headers(req)
+            });
 
         trace!(
             hash = &digest.hash,
@@ -177,7 +274,7 @@ impl RemoteClient for GrpcRemoteClient {
 
         match client
             .update_action_result(UpdateActionResultRequest {
-                instance_name: self.instance_name.clone(),
+                instance_name: self.config.cache.instance_name.clone(),
                 action_digest: Some(digest.to_owned()),
                 action_result: Some(result),
                 digest_function: digest_function::Value::Sha256 as i32,
@@ -210,9 +307,30 @@ impl RemoteClient for GrpcRemoteClient {
 
                     Ok(None)
                 } else {
+                    // dbg!("update_action_result", digest);
                     Err(map_status_error(status).into())
                 }
             }
+        }
+    }
+
+    // https://github.com/bazelbuild/remote-apis/blob/main/build/bazel/remote/execution/v2/remote_execution.proto#L351
+    async fn find_missing_blobs(&self, blob_digests: Vec<Digest>) -> miette::Result<Vec<Digest>> {
+        let mut client = ContentAddressableStorageClient::with_interceptor(
+            self.channel.clone().unwrap(),
+            |req| self.inject_auth_headers(req),
+        );
+
+        match client
+            .find_missing_blobs(FindMissingBlobsRequest {
+                instance_name: self.config.cache.instance_name.clone(),
+                blob_digests,
+                digest_function: digest_function::Value::Sha256 as i32,
+            })
+            .await
+        {
+            Ok(response) => Ok(response.into_inner().missing_blob_digests),
+            Err(status) => Err(map_status_error(status).into()),
         }
     }
 
@@ -222,19 +340,22 @@ impl RemoteClient for GrpcRemoteClient {
         digest: &Digest,
         blob_digests: Vec<Digest>,
     ) -> miette::Result<Vec<Blob>> {
-        let mut client = ContentAddressableStorageClient::new(self.channel.clone().unwrap());
+        let mut client = ContentAddressableStorageClient::with_interceptor(
+            self.channel.clone().unwrap(),
+            |req| self.inject_auth_headers(req),
+        );
 
         trace!(
             hash = &digest.hash,
-            compression = self.compression.to_string(),
+            compression = self.config.cache.compression.to_string(),
             "Downloading {} output blobs",
             blob_digests.len()
         );
 
         let response = match client
             .batch_read_blobs(BatchReadBlobsRequest {
-                acceptable_compressors: get_acceptable_compressors(self.compression),
-                instance_name: self.instance_name.clone(),
+                acceptable_compressors: get_acceptable_compressors(self.config.cache.compression),
+                instance_name: self.config.cache.instance_name.clone(),
                 digests: blob_digests,
                 digest_function: digest_function::Value::Sha256 as i32,
             })
@@ -250,6 +371,7 @@ impl RemoteClient for GrpcRemoteClient {
 
                     Ok(vec![])
                 } else {
+                    // dbg!("batch_read_blobs", digest);
                     Err(map_status_error(status).into())
                 };
             }
@@ -260,11 +382,19 @@ impl RemoteClient for GrpcRemoteClient {
 
         for download in response.into_inner().responses {
             if let Some(status) = download.status {
-                if status.code != 0 {
+                let code = Code::from_i32(status.code);
+
+                if !matches!(code, Code::Ok | Code::NotFound) {
                     warn!(
+                        hash = &digest.hash,
                         details = ?status.details,
+                        code = ?code,
                         "Failed to download blob: {}",
-                        status.message
+                        if status.message.is_empty() {
+                            code.to_string()
+                        } else {
+                            status.message
+                        }
                     );
                 }
             }
@@ -272,7 +402,10 @@ impl RemoteClient for GrpcRemoteClient {
             if let Some(digest) = download.digest {
                 blobs.push(Blob {
                     digest,
-                    bytes: decompress_blob(self.compression, download.data)?,
+                    bytes: decompress_blob(
+                        get_compression_from_code(download.compressor),
+                        download.data,
+                    )?,
                 });
             }
 
@@ -295,11 +428,14 @@ impl RemoteClient for GrpcRemoteClient {
         digest: &Digest,
         blobs: Vec<Blob>,
     ) -> miette::Result<Vec<Option<Digest>>> {
-        let mut client = ContentAddressableStorageClient::new(self.channel.clone().unwrap());
+        let mut client = ContentAddressableStorageClient::with_interceptor(
+            self.channel.clone().unwrap(),
+            |req| self.inject_auth_headers(req),
+        );
 
         trace!(
             hash = &digest.hash,
-            compression = self.compression.to_string(),
+            compression = self.config.cache.compression.to_string(),
             "Uploading {} output blobs",
             blobs.len()
         );
@@ -309,14 +445,14 @@ impl RemoteClient for GrpcRemoteClient {
         for blob in blobs {
             requests.push(batch_update_blobs_request::Request {
                 digest: Some(blob.digest),
-                data: compress_blob(self.compression, blob.bytes)?,
-                compressor: get_compressor(self.compression),
+                data: compress_blob(self.config.cache.compression, blob.bytes)?,
+                compressor: get_compressor(self.config.cache.compression),
             });
         }
 
         let response = match client
             .batch_update_blobs(BatchUpdateBlobsRequest {
-                instance_name: self.instance_name.clone(),
+                instance_name: self.config.cache.instance_name.clone(),
                 requests,
                 digest_function: digest_function::Value::Sha256 as i32,
             })
@@ -342,6 +478,7 @@ impl RemoteClient for GrpcRemoteClient {
 
                     Ok(vec![])
                 } else {
+                    // dbg!("batch_update_blobs", digest);
                     Err(map_status_error(status).into())
                 };
             }
@@ -352,11 +489,19 @@ impl RemoteClient for GrpcRemoteClient {
 
         for upload in response.into_inner().responses {
             if let Some(status) = upload.status {
-                if status.code != 0 {
+                let code = Code::from_i32(status.code);
+
+                if !matches!(code, Code::Ok) {
                     warn!(
+                        hash = &digest.hash,
                         details = ?status.details,
+                        code = ?code,
                         "Failed to upload blob: {}",
-                        status.message
+                        if status.message.is_empty() {
+                            code.to_string()
+                        } else {
+                            status.message
+                        }
                     );
                 }
             }