+++ crates/remote/src/lib.rs
@@ -1,3 +1,4 @@
+mod action_state;
 mod compression;
 mod fs_digest;
 mod grpc_remote_client;
@@ -8,6 +9,7 @@ mod remote_client;
 mod remote_error;
 mod remote_service;
 
+pub use action_state::*;
 pub use bazel_remote_apis::build::bazel::remote::execution::v2::Digest;
 pub use fs_digest::*;
 pub use remote_error::*;