+++ crates/task-runner/tests/output_hydrater_test.rs
@@ -1,22 +1,9 @@
 mod utils;
 
-use moon_action::Operation;
 use moon_cache::CacheMode;
-use moon_remote::Digest;
 use moon_task_runner::output_hydrater::HydrateFrom;
 use std::env;
 use utils::*;
-
-fn stub_digest() -> Digest {
-    Digest {
-        hash: "hash123".into(),
-        size_bytes: 0,
-    }
-}
-
-fn stub_operation() -> Operation {
-    Operation::output_hydration()
-}
 
 mod output_hydrater {
     use super::*;
@@ -36,11 +23,9 @@ mod output_hydrater {
         async fn does_nothing_if_from_prev_outputs() {
             let container = TaskRunnerContainer::new("archive", "file-outputs").await;
             let hydrater = container.create_hydrator();
-            let digest = stub_digest();
-            let mut operation = stub_operation();
 
             assert!(hydrater
-                .hydrate(HydrateFrom::PreviousOutput, &digest, &mut operation)
+                .hydrate(HydrateFrom::PreviousOutput, "hash123", None)
                 .await
                 .unwrap());
         }
@@ -53,8 +38,6 @@ mod output_hydrater {
                 .create_file(".moon/cache/outputs/hash123.tar.gz", "");
 
             let hydrater = container.create_hydrator();
-            let digest = stub_digest();
-            let mut operation = stub_operation();
 
             container
                 .app_context
@@ -62,7 +45,7 @@ mod output_hydrater {
                 .force_mode(CacheMode::Off);
 
             assert!(!hydrater
-                .hydrate(HydrateFrom::LocalCache, &digest, &mut operation)
+                .hydrate(HydrateFrom::LocalCache, "hash123", None)
                 .await
                 .unwrap());
 
@@ -77,8 +60,6 @@ mod output_hydrater {
                 .create_file(".moon/cache/outputs/hash123.tar.gz", "");
 
             let hydrater = container.create_hydrator();
-            let digest = stub_digest();
-            let mut operation = stub_operation();
 
             container
                 .app_context
@@ -86,7 +67,7 @@ mod output_hydrater {
                 .force_mode(CacheMode::Write);
 
             assert!(!hydrater
-                .hydrate(HydrateFrom::LocalCache, &digest, &mut operation)
+                .hydrate(HydrateFrom::LocalCache, "hash123", None)
                 .await
                 .unwrap());
 
@@ -101,11 +82,9 @@ mod output_hydrater {
             assert!(!container.sandbox.path().join("project/file.txt").exists());
 
             let hydrater = container.create_hydrator();
-            let digest = stub_digest();
-            let mut operation = stub_operation();
 
             hydrater
-                .hydrate(HydrateFrom::LocalCache, &digest, &mut operation)
+                .hydrate(HydrateFrom::LocalCache, "hash123", None)
                 .await
                 .unwrap();
 
@@ -124,11 +103,9 @@ mod output_hydrater {
                 .exists());
 
             let hydrater = container.create_hydrator();
-            let digest = stub_digest();
-            let mut operation = stub_operation();
 
             hydrater
-                .hydrate(HydrateFrom::LocalCache, &digest, &mut operation)
+                .hydrate(HydrateFrom::LocalCache, "hash123", None)
                 .await
                 .unwrap();