+++ crates/task-runner/tests/task_runner_test.rs
@@ -3,20 +3,12 @@ mod utils;
 use moon_action::ActionStatus;
 use moon_action_context::*;
 use moon_cache::CacheMode;
-use moon_remote::Digest;
 use moon_task::Target;
 use moon_task_runner::output_hydrater::HydrateFrom;
 use moon_task_runner::TaskRunner;
 use moon_time::now_millis;
 use std::env;
 use utils::*;
-
-fn stub_digest() -> Digest {
-    Digest {
-        hash: "hash123".into(),
-        size_bytes: 0,
-    }
-}
 
 mod task_runner {
     use super::*;
@@ -709,7 +701,6 @@ mod task_runner {
 
         fn setup_exec_state(runner: &mut TaskRunner) {
             runner.report_item.hash = Some("hash123".into());
-            runner.action_digest = stub_digest();
         }
 
         #[tokio::test]
@@ -726,11 +717,7 @@ mod task_runner {
             runner.execute(&context, &node).await.unwrap();
 
             assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
+                runner.target_state.as_ref().unwrap(),
                 &TargetState::Passed("hash123".into())
             );
         }
@@ -747,11 +734,7 @@ mod task_runner {
             runner.execute(&context, &node).await.unwrap();
 
             assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
+                runner.target_state.as_ref().unwrap(),
                 &TargetState::Passthrough
             );
         }
@@ -770,14 +753,7 @@ mod task_runner {
             // Swallow panic so we can check operations
             let _ = runner.execute(&context, &node).await;
 
-            assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
-                &TargetState::Failed
-            );
+            assert_eq!(runner.target_state.as_ref().unwrap(), &TargetState::Failed);
         }
 
         #[tokio::test]
@@ -895,9 +871,8 @@ mod task_runner {
         async fn creates_an_operation() {
             let container = TaskRunnerContainer::new("runner", "base").await;
             let mut runner = container.create_runner();
-            let context = ActionContext::default();
 
-            runner.skip(&context).unwrap();
+            runner.skip().unwrap();
 
             let operation = runner.operations.last().unwrap();
 
@@ -909,18 +884,10 @@ mod task_runner {
         async fn sets_skipped_state() {
             let container = TaskRunnerContainer::new("runner", "base").await;
             let mut runner = container.create_runner();
-            let context = ActionContext::default();
 
-            runner.skip(&context).unwrap();
+            runner.skip().unwrap();
 
-            assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
-                &TargetState::Skipped
-            );
+            assert_eq!(runner.target_state.as_ref().unwrap(), &TargetState::Skipped);
         }
     }
 
@@ -931,9 +898,8 @@ mod task_runner {
         async fn creates_an_operation() {
             let container = TaskRunnerContainer::new("runner", "base").await;
             let mut runner = container.create_runner();
-            let context = ActionContext::default();
 
-            runner.skip_no_op(&context).unwrap();
+            runner.skip_no_op().unwrap();
 
             let operation = runner.operations.last().unwrap();
 
@@ -945,16 +911,11 @@ mod task_runner {
         async fn sets_passthrough_state() {
             let container = TaskRunnerContainer::new("runner", "base").await;
             let mut runner = container.create_runner();
-            let context = ActionContext::default();
 
-            runner.skip_no_op(&context).unwrap();
+            runner.skip_no_op().unwrap();
 
             assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
+                runner.target_state.as_ref().unwrap(),
                 &TargetState::Passthrough
             );
         }
@@ -962,17 +923,12 @@ mod task_runner {
         async fn sets_completed_state() {
             let container = TaskRunnerContainer::new("runner", "base").await;
             let mut runner = container.create_runner();
-            let context = ActionContext::default();
 
             runner.report_item.hash = Some("hash123".into());
-            runner.skip_no_op(&context).unwrap();
+            runner.skip_no_op().unwrap();
 
             assert_eq!(
-                context
-                    .target_states
-                    .get(&runner.task.target)
-                    .unwrap()
-                    .get(),
+                runner.target_state.as_ref().unwrap(),
                 &TargetState::Passed("hash123".into())
             );
         }
@@ -1046,8 +1002,7 @@ mod task_runner {
 
                 let mut runner = container.create_runner();
 
-                let context = ActionContext::default();
-                let result = runner.hydrate(&context, "hash123").await.unwrap();
+                let result = runner.hydrate("hash123").await.unwrap();
 
                 assert!(!result);
 
@@ -1067,7 +1022,6 @@ mod task_runner {
 
                 runner.cache.data.exit_code = 0;
                 runner.cache.data.hash = "hash123".into();
-                runner.action_digest = stub_digest();
             }
 
             #[tokio::test]
@@ -1077,8 +1031,7 @@ mod task_runner {
 
                 setup_previous_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                let result = runner.hydrate(&context, "hash123").await.unwrap();
+                let result = runner.hydrate("hash123").await.unwrap();
 
                 assert!(result);
 
@@ -1095,15 +1048,10 @@ mod task_runner {
 
                 setup_previous_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                runner.hydrate(&context, "hash123").await.unwrap();
+                runner.hydrate("hash123").await.unwrap();
 
                 assert_eq!(
-                    context
-                        .target_states
-                        .get(&runner.task.target)
-                        .unwrap()
-                        .get(),
+                    runner.target_state.as_ref().unwrap(),
                     &TargetState::Passed("hash123".into())
                 );
             }
@@ -1113,11 +1061,9 @@ mod task_runner {
             use super::*;
             use std::fs;
 
-            fn setup_local_state(container: &TaskRunnerContainer, runner: &mut TaskRunner) {
+            fn setup_local_state(container: &TaskRunnerContainer, _runner: &mut TaskRunner) {
                 container.sandbox.enable_git();
                 container.pack_archive();
-
-                runner.action_digest = stub_digest();
             }
 
             #[tokio::test]
@@ -1127,8 +1073,7 @@ mod task_runner {
 
                 setup_local_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                let result = runner.hydrate(&context, "hash123").await.unwrap();
+                let result = runner.hydrate("hash123").await.unwrap();
 
                 assert!(result);
 
@@ -1145,15 +1090,10 @@ mod task_runner {
 
                 setup_local_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                runner.hydrate(&context, "hash123").await.unwrap();
+                runner.hydrate("hash123").await.unwrap();
 
                 assert_eq!(
-                    context
-                        .target_states
-                        .get(&runner.task.target)
-                        .unwrap()
-                        .get(),
+                    runner.target_state.as_ref().unwrap(),
                     &TargetState::Passed("hash123".into())
                 );
             }
@@ -1165,8 +1105,7 @@ mod task_runner {
 
                 setup_local_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                runner.hydrate(&context, "hash123").await.unwrap();
+                runner.hydrate("hash123").await.unwrap();
 
                 let output_file = container.sandbox.path().join("project/file.txt");
 
@@ -1181,8 +1120,7 @@ mod task_runner {
 
                 setup_local_state(&container, &mut runner);
 
-                let context = ActionContext::default();
-                let result = runner.hydrate(&context, "hash123").await.unwrap();
+                let result = runner.hydrate("hash123").await.unwrap();
 
                 assert!(result);